requirements.txt
# ---- Core runtime ----
fastapi==0.115.0
uvicorn[standard]==0.30.1
starlette==0.38.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings>=2.0,<3
typing_extensions==4.15.0
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.3

# ---- Google Ads / Auth ----
google-ads==28.3.0
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
proto-plus==1.26.1
protobuf==6.33.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
rsa==4.9.1

# ---- OpenAI (Responses API) ----
openai==2.6.1
httpx==0.27.2
httpcore==1.0.9
anyio==4.11.0
sniffio==1.3.1
tqdm==4.67.1

# ---- Async / Network helpers ----
aiohttp==3.10.5
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
urllib3==2.5.0
idna==3.11
certifi==2025.10.5
charset-normalizer==3.4.4

# ---- Dev / Testing / Linting ----
black==24.10.0
ruff==0.6.9
pytest==8.3.3
click==8.3.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
mypy_extensions==1.1.0
iniconfig==2.3.0
pluggy==1.6.0
setuptools==80.9.0
wheel==0.45.1

# ---- Optional server performance ----
uvloop==0.22.1
httptools==0.7.1
watchfiles==1.1.1
websockets==15.0.1

# ---- Utilities / misc ----
attrs==25.4.0
cachetools==6.2.1
greenlet==3.2.4
distro==1.9.0
PyYAML==6.0.3
propcache==0.4.1
jiter==0.11.1
git-filter-repo==2.47.0
annotated-doc==0.0.3
annotated-types==0.7.0
